third_party/WebKit/LayoutTests/http/tests/locks/requestLock.html

<!DOCTYPE html>
<title>Web Locks API: navigator.locks.acquire method</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async t => {
  await promise_rejects(t, new TypeError(), navigator.locks.acquire());
  await promise_rejects(t, new TypeError(), navigator.locks.acquire('s'));
}, 'navigator.locks.acquire requires a scope and a callback');

promise_test(async t => {
  await promise_rejects(t, new TypeError(), navigator.locks.acquire('s', lock => {}, {mode: 'foo'}));
  await promise_rejects(t, new TypeError(), navigator.locks.acquire('s', lock => {}, {mode: null }));
}, 'mode must be "shared" or "exclusive"');

promise_test(async t => {
  await promise_rejects(t, new TypeError(), navigator.locks.acquire('s', undefined));
  await promise_rejects(t, new TypeError(), navigator.locks.acquire('s', null));
  await promise_rejects(t, new TypeError(), navigator.locks.acquire('s', 123));
  await promise_rejects(t, new TypeError(), navigator.locks.acquire('s', 'abc'));
  await promise_rejects(t, new TypeError(), navigator.locks.acquire('s', []));
  await promise_rejects(t, new TypeError(), navigator.locks.acquire('s', {}));
}, 'callback must be a function');

promise_test(async t => {
  let release;
  const promise = new Promise(r => { release = r; });

  let returned = navigator.locks.acquire('s', lock => { return promise; });

  const order = [];

  returned.then(() => { order.push('returned'); });
  promise.then(() => { order.push('holding'); });

  release();

  await Promise.all([returned, promise]);

  assert_array_equals(order, ['holding', 'returned']);

}, 'navigator.locks.acquire\'s returned promise resolves after lock is released');

</script>
